Removing useless interfaces from variableInjector and moving to another package
