--- src/main/java/com/browserstack/automate/ci/common/Constants.java
@@ -1,9 +1,0 @@
-package com.browserstack.automate.ci.common;
-
-public class Constants {
-
-  public class VariableInjectorAction {
-    public static final String ICON_FILE_NAME = "document-properties.gif";
-    public static final String DISPLAY_NAME = "Environment Variables";
-  }
-}
